"Ordem de serviço" event type changed
The app was toggling a event for every key pressed and filled the value of Ordem de serviço control, but that was limiting to have just one control that has barcode reader as input device.
Now users need to click on the control he want to fill in and the new event type is change value.